requirements.txt
streamlit>=1.28.0
faster-whisper>=0.10.0
pydub>=0.25.1
python-docx>=1.0.0
Pillow>=10.0.0
genanki>=0.13.0
requests>=2.31.0
